+++ src/base/schema.rs
@@ -74,7 +74,6 @@ pub struct TransactionDataBulkNodeV1 {
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
 /// Set of transactions inside a bulk transaction
 pub struct BulkTransactions {
-    // is box right approach?
     pub root: Box<UnsignedTransaction>,
     pub nodes: Option<Vec<SignedTransaction>>,
 }
@@ -89,13 +88,13 @@ pub struct TransactionDataBulkV1 {
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
 #[serde(tag = "schema")]
 pub enum TransactionData {
-    #[serde(rename = "v1")]
+    #[serde(rename = "d8af8f563f25eb065651ccdd05b9726fd27ff9dc40e3b9c8b4d2c55fa9819f36")]
     V1(TransactionDataV1),
-    #[serde(rename = "bnv1")]
+    #[serde(rename = "097e5f552c79d4f64e15f853ad19d013973343aea557d5c1482d9cef71915db8")]
     BulkNodeV1(TransactionDataBulkNodeV1),
-    #[serde(rename = "brv1")]
+    #[serde(rename = "0bccf5dce4f25036de1ef091ea9e862fa348e6de82ef16fbcfc84c1f1314b86e")]
     BulkRootV1(TransactionDataV1),
-    #[serde(rename = "bv1")]
+    #[serde(rename = "0ec3469e3509682d7599797a9d1c5cdf56b2d9bd435f853a3b999cbb717e0337")]
     BulkV1(TransactionDataBulkV1),
 }
 
@@ -631,12 +630,17 @@ pub struct BlockchainSettings {
     pub block_timeout: u16,
     /// Name of the method in Trinci account to burn fuel
     pub burning_fuel_method: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
+    /// Name of the blockchain network
     pub network_name: Option<String>,
+    /// Flag that identify a production environment
+    pub is_production: bool,
+    /// Compatibility of the boostrap.bin
+    pub min_node_version: String,
 }
 
 #[cfg(test)]
 pub mod tests {
+
     use super::*;
     use crate::{
         base::serialize::MessagePack,
@@ -649,23 +653,24 @@ pub mod tests {
 
     const ACCOUNT_ID: &str = "QmNLei78zWmzUdbeRB3CiUfAizWUrbeeZh5K1rhAQKCh51";
 
-    const UNIT_TRANSACTION_DATA_HEX: &str = "99a27631d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461";
+    const UNIT_TRANSACTION_DATA_HEX: &str = "99d94064386166386635363366323565623036353635316363646430356239373236666432376666396463343065336239633862346432633535666139383139663336d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461";
     const UNIT_TRANSACTION_DATA_HASH_HEX: &str =
-        "1220970572e00cacd21dd115e12ed6809f6dcc52f06cbe6e2a96e5e22b370126cc1b";
+        "12207cfff11a272ad3f5cb60606717adc9984d1cd4dc4c491fdf4c56661ee40caaad";
 
-    const BULK_TRANSACTION_DATA_HEX: &str = "92a3627631929199a462727631d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461c0";
+    const BULK_TRANSACTION_DATA_HEX: &str = "92d94030656333343639653335303936383264373539393739376139643163356364663536623264396264343335663835336133623939396362623731376530333337929199d94030626363663564636534663235303336646531656630393165613965383632666133343865366465383265663136666263666338346331663133313462383665d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461c0";
     const BULK_WITH_NODES_TRANSACTION_DATA_HASH_HEX: &str =
-        "12209d02cba525b07b4208b3bfb8bd71f41e625498c402e9d3e00c7ecbf5e3c649ea";
-    const BULK_WITH_NODES_TRANSACTION_DATA_HEX: &str = "92a3627631929199a462727631d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f70617175654461746192929aa4626e7631d92e516d56513965715a4d33686f5463597a4366365241585078715368583163396a4d48445659444875707862575232cd03e8c403000102a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea86d6574686f645f3193a56563647361a9736563703338347231c461044717583406373a9b47f564e6af4c28d9bc45b11da5de0fdfcd9928dab12eaacaedfabc7357565f2ecfa222f4b4e654a727397c3cad00a2af4c21defe5a0b403d3e62390b71633b203c268fd35ffe2e83fc7c602c2ae19274707a96f579e5439ec403000102c422122000b09762946a4d8f74928eb1b882d0ac2d92803f00199a8af9c89ee411fe54c2c460c7e0cf6099d3cc1f6cd81fcf75c12af8a378c9433299938f15d3ebeee07a9756546420d9a63a83abe4d52dc41b2ab1487d22f54cc5a713daf9cde22341c5fecd24108a73ea2d77bca871e7e16cbe4f07b427db94d4b4ea0051c829b474402d8a929aa4626e7631d92e516d63326d383963624d6d5956516156484c64336655365a6766646d46316d3775537538643246584c35697a745acd03e8c403000102a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea86d6574686f645f3293a56563647361a9736563703338347231c461048295bf9d82b141177ebc03002000573f556bf40bfbd9b70e20c92c45374ed87cdd099759a019dc8665114ab60c83052c37fa99ac6cd933dc635e690d616d42874114dce3b1f41f5fc9b563679e50b0cb5f9295ce8fbaaa812ef69fef6b9b6100c403000102c422122000b09762946a4d8f74928eb1b882d0ac2d92803f00199a8af9c89ee411fe54c2c4605365838d6ea99075b7b08c9186ddd28656b18fa9e975acb5f1bc48b99f46848044a2ab506ebf5da5cdc9e9bd4525751cb537eb30d5d984aca6bd30c2416363b26ce25ca78c33091e7482d150777a69dd082a9b68656c4bc6b5fdd198e86ece53";
+        "1220cb7525e6f80b116271e6fc4bbf99b3a10815b3380fc32be2c990a0b2c547bbad";
+    const BULK_WITH_NODES_TRANSACTION_DATA_HEX: &str = "92d94030656333343639653335303936383264373539393739376139643163356364663536623264396264343335663835336133623939396362623731376530333337929199d94030626363663564636534663235303336646531656630393165613965383632666133343865366465383265663136666263666338346331663133313462383665d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f70617175654461746192929ad94030393765356635353263373964346636346531356638353361643139643031333937333334336165613535376435633134383264396365663731393135646238d92e516d56513965715a4d33686f5463597a4366365241585078715368583163396a4d48445659444875707862575232cd03e8c403000102a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea86d6574686f645f3193a56563647361a9736563703338347231c461044717583406373a9b47f564e6af4c28d9bc45b11da5de0fdfcd9928dab12eaacaedfabc7357565f2ecfa222f4b4e654a727397c3cad00a2af4c21defe5a0b403d3e62390b71633b203c268fd35ffe2e83fc7c602c2ae19274707a96f579e5439ec403000102c4221220bae7ad2ad1a246565d642930c241e7c390b159878cfa5a1a76e4fb91a00595ddc460c7e0cf6099d3cc1f6cd81fcf75c12af8a378c9433299938f15d3ebeee07a9756546420d9a63a83abe4d52dc41b2ab1487d22f54cc5a713daf9cde22341c5fecd24108a73ea2d77bca871e7e16cbe4f07b427db94d4b4ea0051c829b474402d8a929ad94030393765356635353263373964346636346531356638353361643139643031333937333334336165613535376435633134383264396365663731393135646238d92e516d63326d383963624d6d5956516156484c64336655365a6766646d46316d3775537538643246584c35697a745acd03e8c403000102a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea86d6574686f645f3293a56563647361a9736563703338347231c461048295bf9d82b141177ebc03002000573f556bf40bfbd9b70e20c92c45374ed87cdd099759a019dc8665114ab60c83052c37fa99ac6cd933dc635e690d616d42874114dce3b1f41f5fc9b563679e50b0cb5f9295ce8fbaaa812ef69fef6b9b6100c403000102c4221220bae7ad2ad1a246565d642930c241e7c390b159878cfa5a1a76e4fb91a00595ddc4605365838d6ea99075b7b08c9186ddd28656b18fa9e975acb5f1bc48b99f46848044a2ab506ebf5da5cdc9e9bd4525751cb537eb30d5d984aca6bd30c2416363b26ce25ca78c33091e7482d150777a69dd082a9b68656c4bc6b5fdd198e86ece53";
     const BULK_TRANSACTION_DATA_HASH_HEX: &str =
-        "12209d02cba525b07b4208b3bfb8bd71f41e625498c402e9d3e00c7ecbf5e3c649ea";
+        "1220cb7525e6f80b116271e6fc4bbf99b3a10815b3380fc32be2c990a0b2c547bbad";
 
-    const UNIT_TRANSACTION_HEX: &str = "93a7756e69745f747899a27631d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461c46046d493f0edae8a7fdde770f283c26ab0896d93134770be5f7094a53c8d9638f88e140933222da359677ca5abc046e15c30d6abae9370836d86579e1a4717b285e90e87226ced0d9668cf4fd44c13078c7f8bf77d21ddddb72b123bb73080155c";
-    const UNIT_TRANSACTION_SIGN: &str = "46d493f0edae8a7fdde770f283c26ab0896d93134770be5f7094a53c8d9638f88e140933222da359677ca5abc046e15c30d6abae9370836d86579e1a4717b285e90e87226ced0d9668cf4fd44c13078c7f8bf77d21ddddb72b123bb73080155c";
+    const UNIT_TRANSACTION_HEX: &str = "93a7756e69745f747899d94064386166386635363366323565623036353635316363646430356239373236666432376666396463343065336239633862346432633535666139383139663336d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461c4603b472dc456b2db807f13b60920ee0c663405441b3b5e46f93371e40ecbe1a7922034865a4be36812394e7f65b26e626737c125174889888d30a85e131b3a75416cfa361621731853325b8383afd283beaaf244ebd9f86b88d8b275376f428fdb";
+    const UNIT_TRANSACTION_SIGN: &str = "3b472dc456b2db807f13b60920ee0c663405441b3b5e46f93371e40ecbe1a7922034865a4be36812394e7f65b26e626737c125174889888d30a85e131b3a75416cfa361621731853325b8383afd283beaaf244ebd9f86b88d8b275376f428fdb";
 
-    const BULK_WITHOUT_NODES_TRANSACTION_HEX: &str = "93a762756c6b5f747892a3627631929199a462727631d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461c0c460a44e61159c7d9171fadbf2e6e14f5082466129456ee4b19c17c9f2d814e7aeb91697f8650ffd3e1577c90142a310d93029fc30a4eab21c823b2ae1b7ac39dd6ea9778062a936bc7aad4e04427c3375b91fc8bf989c165bf116e26d3bb1d0e7c8";
-    const BULK_WITHOUT_NODES_TRANSACTION_SIGN: &str = "a44e61159c7d9171fadbf2e6e14f5082466129456ee4b19c17c9f2d814e7aeb91697f8650ffd3e1577c90142a310d93029fc30a4eab21c823b2ae1b7ac39dd6ea9778062a936bc7aad4e04427c3375b91fc8bf989c165bf116e26d3bb1d0e7c8";
-    const BULK_WITH_NODES_TRANSACTION_SIGN: &str = "52abc48510a10737c3ddf49de14474f232bfb81bc0f3d32be2ef7f390b52d025a561156f411b67f4e96d4634ee99ef6929c9ad636de7eabaacc2499234bc7f3baec3a2b9c542958a9c2557a4f136dbba824ddc65a04e1dc0447d27f0790c407a";
+    const BULK_WITHOUT_NODES_TRANSACTION_HEX: &str = "93a762756c6b5f747892d94030656333343639653335303936383264373539393739376139643163356364663536623264396264343335663835336133623939396362623731376530333337929199d94030626363663564636534663235303336646531656630393165613965383632666133343865366465383265663136666263666338346331663133313462383665d92e516d59486e45514c64663568374b59626a4650754853526b325350676458724a5746683557363936485066713769cd03e8c408ab82b741e023a412a6736b796e6574c42212202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7aea97465726d696e61746593a56563647361a9736563703338347231c461045936d631b849bb5760bcf62e0d1261b6b6e227dc0a3892cbeec91be069aaa25996f276b271c2c53cba4be96d67edcadd66b793456290609102d5401f413cd1b5f4130b9cfaa68d30d0d25c3704cb72734cd32064365ff7042f5a3eee09b06cc1c40a4f706171756544617461c0c46019f3a5d479c463569346078e52bc57d76201232eaaea083b54777086732e5f4a47bc3d8eee4e5af46eee211e40fac6502a1a0710161865a6abd13f1528ff0e4611d3af2d0b89e57ce3b2af0d283252aec2ecca5c5317cf7b98bdad96c9099299";
+    const BULK_WITHOUT_NODES_TRANSACTION_SIGN: &str = "19f3a5d479c463569346078e52bc57d76201232eaaea083b54777086732e5f4a47bc3d8eee4e5af46eee211e40fac6502a1a0710161865a6abd13f1528ff0e4611d3af2d0b89e57ce3b2af0d283252aec2ecca5c5317cf7b98bdad96c9099299";
+    const BULK_WITH_NODES_TRANSACTION_SIGN_ALT: &str = "c814893010201f45925d4118fdc5beabc7cb1d6cf58c9e44521a31f01d62ee445bf15729b304f9b32f4e6531c092f3b3fa3f893504dbd18c5f39c6ebb5a16a0584fd811dd43ca661dba909f0ee75965d8f6642a76b1c254514166f0cd4a78269";
+    const BULK_WITH_NODES_TRANSACTION_SIGN: &str = "b0315a41eac7178b2b5a06b8faf9e19647086e34c789a0fbc87134350392b093942cf5de7a557a23c8397b60fb0a5c14a3970beac5db1c30bd97caa4e63c99861b5db79769f6e497d28b4cd1f862d7512f6f7702810f9e361386db6c91158d84";
 
     const RECEIPT_HEX: &str = "960309cd03e7c3c40a4f70617175654461746190";
     const RECEIPT_HASH_HEX: &str =
@@ -726,7 +731,11 @@ pub mod tests {
         })
     }
 
-    fn create_bulk_node_transaction(root_hash: Hash, key: u8, method: String) -> SignedTransaction {
+    fn create_bulk_node_transaction_alt(
+        root_hash: Hash,
+        key: u8,
+        method: String,
+    ) -> SignedTransaction {
         let keypair = KeyPair::Ecdsa(ecdsa_secp384_test_keypair(key));
         let contract =
             Hash::from_hex("12202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae")
@@ -741,6 +750,26 @@ pub mod tests {
             2 => hex::decode("5365838d6ea99075b7b08c9186ddd28656b18fa9e975acb5f1bc48b99f46848044a2ab506ebf5da5cdc9e9bd4525751cb537eb30d5d984aca6bd30c2416363b26ce25ca78c33091e7482d150777a69dd082a9b68656c4bc6b5fdd198e86ece53").unwrap(),
             _ => panic!()
         };
+
+        SignedTransaction { data, signature }
+    }
+
+    fn create_bulk_node_transaction(root_hash: Hash, key: u8, method: String) -> SignedTransaction {
+        let keypair = KeyPair::Ecdsa(ecdsa_secp384_test_keypair(key));
+        let contract =
+            Hash::from_hex("12202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae")
+                .unwrap();
+
+        let data =
+            create_transactiondata_bulk_node_v1(keypair.public_key(), contract, method, root_hash);
+
+        let signature = data.sign(&keypair).unwrap();
+        //let signature = match key{
+        //    // This guarantees a fix signature for tests
+        //    1 => hex::decode("c7e0cf6099d3cc1f6cd81fcf75c12af8a378c9433299938f15d3ebeee07a9756546420d9a63a83abe4d52dc41b2ab1487d22f54cc5a713daf9cde22341c5fecd24108a73ea2d77bca871e7e16cbe4f07b427db94d4b4ea0051c829b474402d8a").unwrap(),
+        //    2 => hex::decode("5365838d6ea99075b7b08c9186ddd28656b18fa9e975acb5f1bc48b99f46848044a2ab506ebf5da5cdc9e9bd4525751cb537eb30d5d984aca6bd30c2416363b26ce25ca78c33091e7482d150777a69dd082a9b68656c4bc6b5fdd198e86ece53").unwrap(),
+        //    _ => panic!()
+        //};
 
         SignedTransaction { data, signature }
     }
@@ -770,10 +799,55 @@ pub mod tests {
         let root_hash = root.data.primary_hash();
 
         let nodes = if with_nodes {
-            Some(vec![
+            let nodes = vec![
                 create_bulk_node_transaction(root_hash, 1, String::from("method_1")),
                 create_bulk_node_transaction(root_hash, 2, String::from("method_2")),
-            ])
+            ];
+
+            Some(nodes)
+        } else {
+            None
+        };
+
+        TransactionData::BulkV1(TransactionDataBulkV1 {
+            txs: BulkTransactions {
+                root: Box::new(root),
+                nodes,
+            },
+        })
+    }
+
+    fn create_test_data_bulk_alt(with_nodes: bool) -> TransactionData {
+        // Opaque information returned by the smart contract.
+        let args = hex::decode("4f706171756544617461").unwrap();
+        let public_key = PublicKey::Ecdsa(ecdsa_secp384_test_public_key(0));
+        let account = public_key.to_account_id();
+        let contract =
+            Hash::from_hex("12202c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae")
+                .unwrap();
+
+        let root_data = TransactionData::BulkRootV1(TransactionDataV1 {
+            account,
+            fuel_limit: FUEL_LIMIT,
+            nonce: [0xab, 0x82, 0xb7, 0x41, 0xe0, 0x23, 0xa4, 0x12].to_vec(),
+            network: "skynet".to_string(),
+            contract: Some(contract),
+            method: "terminate".to_string(),
+            caller: public_key,
+            args,
+        });
+
+        let root = UnsignedTransaction { data: root_data };
+
+        let root_hash = root.data.primary_hash();
+
+        let nodes = if with_nodes {
+            let nodes = vec![
+                create_bulk_node_transaction_alt(root_hash, 1, String::from("method_1")),
+                create_bulk_node_transaction_alt(root_hash, 2, String::from("method_2")),
+            ];
+
+            Some(nodes)
         } else {
             None
         };
@@ -788,11 +862,11 @@ pub mod tests {
 
     pub fn create_test_unit_tx(fuel_limit: u64) -> Transaction {
         // UNCOMMENT THIS to create a new signature
-        // let keypair = crate::crypto::sign::tests::create_test_keypair();
-        // let data = create_test_data_unit();
-        // let data = crate::base::serialize::rmp_serialize(&data).unwrap();
-        // let signature = keypair.sign(&data).unwrap();
-        // println!("unit_sign: {}", hex::encode(&signature));
+        //let keypair = crate::crypto::sign::tests::create_test_keypair();
+        //let data = create_test_data_unit();
+        //let data = crate::base::serialize::rmp_serialize(&data).unwrap();
+        //let signature = keypair.sign(&data).unwrap();
+        //println!("unit_sign: {}", hex::encode(&signature));
 
         let signature = hex::decode(UNIT_TRANSACTION_SIGN).unwrap();
 
@@ -804,11 +878,11 @@ pub mod tests {
 
     pub fn create_test_bulk_tx(with_nodes: bool) -> Transaction {
         // UNCOMMENT THIS to create a new signature
-        // let keypair = crate::crypto::sign::tests::create_test_keypair();
-        // let data = create_test_data_bulk(with_nodes);
-        // let data = crate::base::serialize::rmp_serialize(&data).unwrap();
-        // let signature = keypair.sign(&data).unwrap();
-        // println!("bulk_sign: {}", hex::encode(&signature));
+        //let keypair = crate::crypto::sign::tests::create_test_keypair();
+        //let data = create_test_data_bulk(with_nodes);
+        //let data = crate::base::serialize::rmp_serialize(&data).unwrap();
+        //let signature = keypair.sign(&data).unwrap();
+        //println!("bulk_sign: {}", hex::encode(&signature));
 
         let signature = if with_nodes {
             hex::decode(BULK_WITH_NODES_TRANSACTION_SIGN).unwrap()
@@ -817,6 +891,26 @@ pub mod tests {
         };
         Transaction::BulkTransaction(BulkTransaction {
             data: create_test_data_bulk(with_nodes),
+            signature,
+        })
+    }
+
+    // fixed sign in sub-tx
+    pub fn create_test_bulk_tx_alt(with_nodes: bool) -> Transaction {
+        // UNCOMMENT THIS to create a new signature
+        //let keypair = crate::crypto::sign::tests::create_test_keypair();
+        //let data = create_test_data_bulk_alt(with_nodes);
+        //let data = crate::base::serialize::rmp_serialize(&data).unwrap();
+        //let signature = keypair.sign(&data).unwrap();
+        //println!("bulk_sign: {}", hex::encode(&signature));
+
+        let signature = if with_nodes {
+            hex::decode(BULK_WITH_NODES_TRANSACTION_SIGN_ALT).unwrap()
+        } else {
+            hex::decode(BULK_WITHOUT_NODES_TRANSACTION_SIGN).unwrap()
+        };
+        Transaction::BulkTransaction(BulkTransaction {
+            data: create_test_data_bulk_alt(with_nodes),
             signature,
         })
     }
@@ -939,7 +1033,7 @@ pub mod tests {
 
     #[test]
     fn bulk_with_nodes_transaction_data_serialize() {
-        let data = create_test_data_bulk(true);
+        let data = create_test_data_bulk_alt(true);
 
         let buf = data.serialize();
 
@@ -948,7 +1042,7 @@ pub mod tests {
 
     #[test]
     fn bulk_with_nodes_transaction_data_deserialize() {
-        let expected = create_test_data_bulk(true);
+        let expected = create_test_data_bulk_alt(true);
 
         let buf = hex::decode(BULK_WITH_NODES_TRANSACTION_DATA_HEX).unwrap();
 
@@ -1054,7 +1148,7 @@ pub mod tests {
 
     #[test]
     fn bulk_with_nodes_transaction_data_hash() {
-        let tx = create_test_bulk_tx(false);
+        let tx = create_test_bulk_tx_alt(false);
         let hash = match tx {
             Transaction::BulkTransaction(tx) => tx.data.primary_hash(),
             _ => panic!(),
@@ -1078,7 +1172,7 @@ pub mod tests {
 
     #[test]
     fn bulk_with_nodes_transaction_data_verify() {
-        let tx = create_test_bulk_tx(true);
+        let tx = create_test_bulk_tx_alt(true);
         let result = tx.verify(tx.get_caller(), tx.get_signature());
         assert!(result.is_ok());
     }