+++ src/wm/local.rs
@@ -21,18 +21,19 @@ use crate::{
         schema::SmartContractEvent,
         serialize::{self, rmp_serialize},
     },
-    crypto::Hash,
+    crypto::{drand::SeedSource, Hash},
     db::*,
     wm::{
         host_func::{self, CallContext},
         AppOutput, Wm,
     },
-    Account, Error, ErrorKind, Result,
+    Account, Error, ErrorKind, Result, SERVICE_ACCOUNT_ID,
 };
 use serialize::rmp_deserialize;
 use std::{
     collections::HashMap,
     slice,
+    sync::Arc,
     time::{SystemTime, UNIX_EPOCH},
 };
 use wasmtime::{
@@ -226,6 +227,28 @@ mod local_host_func {
         return_buf(caller, mem, buf)
     }
 
+    /// Check if an account has a method
+    fn is_callable(
+        mut caller: Caller<'_, CallContext>,
+        account_offset: i32,
+        account_size: i32,
+        method_offset: i32,
+        method_size: i32,
+    ) -> std::result::Result<i32, Trap> {
+        // Recover parameters from wasm memory.
+        let mem: Memory = mem_from(&mut caller)?;
+        let buf = slice_from(&mut caller, &mem, account_offset, account_size)?;
+        let account = std::str::from_utf8(buf).map_err(|_| Trap::new("invalid utf-8"))?;
+        let buf = slice_from(&mut caller, &mem, method_offset, method_size)?;
+        let method = std::str::from_utf8(buf).map_err(|_| Trap::new("invalid utf-8"))?;
+        // Recover execution context.
+        let ctx = caller.data_mut();
+        // Invoke portable host function.
+        Ok(host_func::is_callable(ctx, account, method))
+    }
+
+    /// Store asset
+
     fn remove_data(
         mut caller: Caller<'_, CallContext>,
         key_offset: i32,
@@ -277,6 +300,28 @@ mod local_host_func {
         // Invoke portable host function.
         let value = host_func::load_asset(ctx, &account_id);
         return_buf(caller, mem, value)
+    }
+
+    /// Get contract hash from an account
+    fn get_account_contract(
+        mut caller: Caller<'_, CallContext>,
+        account_id_offset: i32,
+        account_id_length: i32,
+    ) -> std::result::Result<WasmSlice, Trap> {
+        // Recover parameters from wasm memory.
+        let mem: Memory = mem_from(&mut caller)?;
+        let buf = slice_from(&mut caller, &mem, account_id_offset, account_id_length)?;
+        let account_id = String::from_utf8_lossy(buf);
+
+        // Recover execution context.
+        let ctx = caller.data_mut();
+
+        let hash = match host_func::get_account_contract(ctx, &account_id) {
+            Some(hash) => hash.as_bytes().to_vec(),
+            None => vec![],
+        };
+
+        return_buf(caller, mem, hash)
     }
 
     /// Digital signature verification.
@@ -351,6 +396,15 @@ mod local_host_func {
         return_buf(caller, mem, hash)
     }
 
+    /// Generate a pseudo random number deterministically, based on the seed
+    fn drand(mut caller: Caller<'_, CallContext>, max: u64) -> std::result::Result<u64, Trap> {
+        // seed from ctx
+        // Recover execution context.
+        let ctx = caller.data_mut();
+        // Invoke portable host fuction
+        Ok(host_func::drand(ctx, max))
+    }
+
     /// Register the required host functions using the same order as the wasm imports list.
     pub(crate) fn host_functions_register(
         mut store: &mut Store<CallContext>,
@@ -366,12 +420,15 @@ mod local_host_func {
                 "hf_load_data" => Func::wrap(&mut store, load_data),
                 "hf_store_data" => Func::wrap(&mut store, store_data),
                 "hf_remove_data" => Func::wrap(&mut store, remove_data),
+                "hf_is_callable" => Func::wrap(&mut store, is_callable),
                 "hf_load_asset" => Func::wrap(&mut store, load_asset),
                 "hf_store_asset" => Func::wrap(&mut store, store_asset),
+                "hf_get_account_contract" => Func::wrap(&mut store, get_account_contract),
                 "hf_get_keys" => Func::wrap(&mut store, get_keys),
                 "hf_call" => Func::wrap(&mut store, call),
                 "hf_verify" => Func::wrap(&mut store, verify),
                 "hf_sha256" => Func::wrap(&mut store, sha256),
+                "hf_drand" => Func::wrap(&mut store, drand),
                 _ => {
                     return Err(Error::new_ext(
                         ErrorKind::NotImplemented,
@@ -396,17 +453,10 @@ struct CachedModule {
     last_used: u64,
 }
 
-/// Closure trait to load a wasm binary.
-pub trait WasmLoader: FnMut(Hash) -> Result<Vec<u8>> + Send + 'static {}
-
-impl<T: FnMut(Hash) -> Result<Vec<u8>> + Send + 'static> WasmLoader for T {}
-
 /// WebAssembly machine using wasmtime as the engine.
 pub struct WmLocal {
     /// Global wasmtime context for compilation and management of wasm modules.
     engine: Engine,
-    /// Callback to load wasm code.
-    loader: Box<dyn WasmLoader>,
     /// Cached wasm modules, ready to be executed.
     cache: HashMap<Hash, CachedModule>,
     /// Maximum cache size.
@@ -421,7 +471,7 @@ impl WmLocal {
     /// Panics if the `cache_max` parameter is zero or if for any reason the
     /// backend fails to initialize.  Failure details are given in the `panic`
     /// error string.
-    pub fn new(loader: impl WasmLoader, cache_max: usize) -> Self {
+    pub fn new(cache_max: usize) -> Self {
         assert!(
             !(cache_max == 0),
             "Fatal: Wm cache size shall be greater than 0"
@@ -432,7 +482,6 @@ impl WmLocal {
 
         WmLocal {
             engine: Engine::new(&config).expect("wm engine creation"),
-            loader: Box::new(loader),
             cache: HashMap::new(),
             cache_max,
         }
@@ -441,7 +490,7 @@ impl WmLocal {
     /// Caches a wasm using the user-provided callback.
     /// If the cache max size has been reached, it removes the least recently
     /// used module from the cache.
-    fn load_module(&mut self, target: &Hash) -> Result<()> {
+    fn load_module(&mut self, engine: &Engine, db: &dyn DbFork, target: &Hash) -> Result<()> {
         let len = self.cache.len();
         if len > self.cache_max {
             let mut iter = self.cache.iter();
@@ -456,10 +505,22 @@ impl WmLocal {
             self.cache.remove(&older_hash);
         }
 
-        let wasm_bin = (self.loader)(*target)?;
+        // let wasm_bin = (self.loader)(db, *target)?;
+        let mut key = String::from("contracts:code:");
+        key.push_str(&hex::encode(&target));
 
-        let module = Module::new(&self.engine, &wasm_bin)
-            .map_err(|err| Error::new_ext(ErrorKind::Other, err))?;
+        let wasm_bin = match db.load_account_data(SERVICE_ACCOUNT_ID, &key) {
+            Some(bin) => bin,
+            None => {
+                return Err(Error::new_ext(
+                    ErrorKind::ResourceNotFound,
+                    "smart contract not found",
+                ))
+            }
+        };
+
+        let module =
+            Module::new(engine, &wasm_bin).map_err(|err| Error::new_ext(ErrorKind::Other, err))?;
 
         let entry = CachedModule {
             module,
@@ -471,9 +532,9 @@ impl WmLocal {
     }
 
     /// Get smart contract module instance from the cache.
-    fn get_module(&mut self, target: &Hash) -> Result<&Module> {
+    fn get_module(&mut self, engine: &Engine, db: &dyn DbFork, target: &Hash) -> Result<&Module> {
         if !self.cache.contains_key(target) {
-            self.load_module(target)?;
+            self.load_module(engine, db, target)?;
         }
         // This should not fail
         let mut entry = self
@@ -486,13 +547,6 @@ impl WmLocal {
             .expect("read system time")
             .as_secs();
         Ok(&entry.module)
-    }
-
-    /// Allow to swap the loader after initialization.
-    /// Loader switch flushes the cache.
-    pub fn set_loader(&mut self, loader: impl WasmLoader) {
-        self.loader = Box::new(loader);
-        self.cache.clear();
     }
 }
 
@@ -579,6 +633,24 @@ impl Wm for WmLocal {
     ) -> Result<Vec<u8>> {
         let app_hash = app_hash_check(db, owner, contract)?;
 
+        let nw_name = String::from("nw_name_test");
+        let nonce: Vec<u8> = vec![0x12, 0x34, 0x56, 0x78, 0x90, 0x12, 0x34, 0x56];
+        let prev_hash =
+            Hash::from_hex("1220a4cea0f0f6eddc6865fd6092a319ccc6d2387cd8bb65e64bdc486f1a9a998569")
+                .unwrap();
+        let txs_hash =
+            Hash::from_hex("1220a4cea0f1f6eddc6865fd6092a319ccc6d2387cf8bb63e64b4c48601a9a998569")
+                .unwrap();
+        let rxs_hash =
+            Hash::from_hex("1220a4cea0f0f6edd46865fd6092a319ccc6d5387cd8bb65e64bdc486f1a9a998569")
+                .unwrap();
+        let seed = SeedSource::new(nw_name, nonce, prev_hash, txs_hash, rxs_hash);
+        let seed = Arc::new(seed);
+
+        let engine1 = self.engine.clone(); // FIXME
+        let engine2 = self.engine.clone();
+        let module = self.get_module(&engine1, db, &app_hash)?;
+
         // Prepare and set execution context for host functions.
         let ctx = CallContext {
             wm: None,
@@ -589,13 +661,11 @@ impl Wm for WmLocal {
             network,
             origin,
             events,
+            seed,
         };
 
         // Allocate execution context (aka Store).
-        let mut store: Store<CallContext> = Store::new(&self.engine, ctx);
-
-        // Get the requested wasm module instance.
-        let module = self.get_module(&app_hash)?;
+        let mut store: Store<CallContext> = Store::new(&engine2, ctx);
 
         // Get imported host functions list.
         let imports = local_host_func::host_functions_register(&mut store, module)?;
@@ -753,16 +823,6 @@ mod tests {
         }
     }
 
-    fn wasm_loader(hash: Hash) -> Result<Vec<u8>> {
-        if hash != test_contract_hash() {
-            return Err(Error::new_ext(
-                ErrorKind::ResourceNotFound,
-                "wasm module not found",
-            ));
-        }
-        Ok(TEST_WASM.to_owned())
-    }
-
     fn create_test_db() -> MockDbFork {
         let mut db = MockDbFork::new();
         db.expect_load_account().returning(|id| {
@@ -777,6 +837,13 @@ mod tests {
         });
         db.expect_store_account().returning(move |_id| ());
         db.expect_state_hash().returning(|_| Hash::default());
+        db.expect_load_account_data().returning(|_, key| {
+            if key == "contracts:code:12201810298b95a12ec9cde9210a81f2a7a5f0e4780da8d4a19b3b8346c0c684e12f"
+            {
+                return None;
+            }
+            Some(TEST_WASM.to_vec())
+        });
         db
     }
 
@@ -820,17 +887,20 @@ mod tests {
 
     #[test]
     fn instance_machine() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let hash = test_contract_hash();
+        let db = create_test_db();
 
-        let result = vm.get_module(&hash);
+        let engine = vm.engine.clone();
+
+        let result = vm.get_module(&engine, &db, &hash);
 
         assert!(result.is_ok());
     }
 
     #[test]
     fn exec_transfer() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data_transfer();
         let mut db = create_test_db();
 
@@ -842,7 +912,7 @@ mod tests {
 
     #[test]
     fn exec_balance() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data_balance();
         let mut db = create_test_db();
 
@@ -854,7 +924,7 @@ mod tests {
 
     #[test]
     fn exec_balance_cached() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data_balance();
         let mut db = create_test_db();
 
@@ -868,7 +938,7 @@ mod tests {
 
     #[test]
     fn exec_inexistent_method() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let args = value!({});
         let data = create_test_data("inexistent", args);
         let mut db = create_test_db();
@@ -884,7 +954,7 @@ mod tests {
 
     #[test]
     fn load_not_existing_module() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let mut data = create_test_data_transfer();
         data.set_contract(Some(Hash::from_hex(NOT_EXISTING_TARGET_HASH).unwrap()));
         data.set_account("NotExistingTestId".to_string());
@@ -895,13 +965,13 @@ mod tests {
         assert_eq!(err.kind, ErrorKind::ResourceNotFound);
         assert_eq!(
             err.to_string_full(),
-            "resource not found: wasm module not found"
+            "resource not found: smart contract not found"
         );
     }
 
     #[test]
     fn echo_generic() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let mut db = create_test_db();
         let input = value!({
             "name": "Davide",
@@ -924,7 +994,7 @@ mod tests {
 
     #[test]
     fn echo_typed() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let mut db = create_test_db();
         let input = value!({
             "name": "Davide",
@@ -947,7 +1017,7 @@ mod tests {
 
     #[test]
     fn echo_typed_bad() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let mut db = create_test_db();
         let input = value!({
             "name": "Davide"
@@ -964,7 +1034,7 @@ mod tests {
 
     #[test]
     fn notify() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let mut db = create_test_db();
         let input = value!({
             "name": "Davide",
@@ -981,7 +1051,7 @@ mod tests {
         assert_eq!(events.len(), 2);
         let event = events.get(0).unwrap();
 
-        assert_eq!(event.event_name, data.get_method());
+        assert_eq!(event.event_name, "event_a");
         assert_eq!(event.emitter_account, data.get_account());
 
         let buf = &event.event_data;
@@ -999,7 +1069,7 @@ mod tests {
 
     #[test]
     fn nested_call() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let mut db = create_test_db();
         let input = value!({
             "name": "Davide"
@@ -1014,7 +1084,7 @@ mod tests {
 
     #[test]
     fn wasm_divide_by_zero() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_data_divide_by_zero();
         let mut db = create_test_db();
 
@@ -1027,7 +1097,7 @@ mod tests {
 
     #[test]
     fn wasm_trigger_panic() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("trigger_panic", value!(null));
         let mut db = create_test_db();
 
@@ -1040,7 +1110,7 @@ mod tests {
 
     #[test]
     fn wasm_exhaust_memory() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("exhaust_memory", value!(null));
         let mut db = create_test_db();
 
@@ -1053,7 +1123,7 @@ mod tests {
 
     #[test]
     fn wasm_infinite_recursion() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("infinite_recursion", value!(true));
         let mut db = create_test_db();
 
@@ -1069,7 +1139,7 @@ mod tests {
 
     #[test]
     fn get_random_sequence() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("get_random_sequence", value!({}));
         let mut db = create_test_db();
 
@@ -1086,7 +1156,7 @@ mod tests {
 
     #[test]
     fn get_hashmap() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("get_hashmap", value!({}));
         let mut db = create_test_db();
 
@@ -1105,7 +1175,7 @@ mod tests {
     #[test]
     #[ignore = "TODO"]
     fn wasm_infinite_loop() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("infinite_loop", value!(null));
         let mut db = create_test_db();
 
@@ -1118,7 +1188,7 @@ mod tests {
 
     #[test]
     fn wasm_null_pointer_indirection() {
-        let mut vm = WmLocal::new(wasm_loader, CACHE_MAX);
+        let mut vm = WmLocal::new(CACHE_MAX);
         let data = create_test_data("null_pointer_indirection", value!(null));
         let mut db = create_test_db();
 
@@ -1126,7 +1196,7 @@ mod tests {
 
         let err_str = err.to_string_full();
         let err_str = err_str.split_inclusive("unreachable").next().unwrap();
-        assert_eq!(err_str, "wasm machine fault: wasm trap: unreachable");
+        assert_eq!(err_str, "wasm machine fault: wasm trap: wasm `unreachable");
     }
 
     #[test]
@@ -1150,4 +1220,6 @@ mod tests {
             println!(">> {:#?}", imp);
         }
     }
+
+    // // TODO: add drand test
 }