+++ CHANGELOG.md
@@ -26,7 +26,26 @@ Additional labels for pre-release metadata:
 * beta.x: shipped version under testing.
 * rc.x: stable release candidate.
 
-0.2.4 - unreleased
+0.2.7 - unreleased
+------------------
+Changed
+* BlockchainSettings struct
+
+0.2.6 - 08-02-2022
+------------------
+Changed
+* removed wasm loader from closure
+
+0.2.5 - 02-02-2022
+------------------
+
+Added
+* Drand implementation
+* get_account_contract host function
+* is_callable host function
+* test mode to p2p module (prevent it from start)
+
+0.2.4 - release skipped
 ------------------
 
 Changed